wip: new structure for Single Combobox
