+++ src/CUIAutoComplete.tsx
@@ -1,0 +1,9 @@
+import React, { FC } from 'react'
+
+type Props = {
+  //
+}
+
+export const CUIAutoComplete: FC<Props> = () => {
+  return <div>WIP</div>
+}
